requirements.txt
discord.py>=2.3.0
yt-dlp[default]>=2024.1.0
aiohttp>=3.9.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0